create component for senders panel
