+++ frontend/src/components/phones/dashboard/Dashboard.tsx
@@ -7,27 +7,20 @@ import {
   CopyIcon,
   ForwardIcon,
   BlockIcon,
-  ChevronLeftIcon,
   ChevronRightIcon,
-  ForwardedCallIcon,
-  ForwardedTextIcon,
-  WarningFilledIcon,
 } from "../../../components/Icons";
-import disabledSendersDataIllustration from "./images/sender-data-disabled-illustration.svg";
 import { MouseEventHandler, useState } from "react";
 import { useRealPhonesData } from "../../../hooks/api/realPhone";
 import { useLocalization } from "@fluent/react";
 import { useInboundContact } from "../../../hooks/api/inboundContact";
-import moment from "moment";
 import { useProfiles } from "../../../hooks/api/profile";
-import { OutboundLink } from "react-ga";
 import { useRuntimeData } from "../../../hooks/api/runtimeData";
+import { SendersPanelView } from "./SendersPanelView";
 
 export const PhoneDashboard = () => {
   const { l10n } = useLocalization();
 
   const runtimeData = useRuntimeData();
-
   const profileData = useProfiles();
   const relayNumber = useRelayNumber();
   const realPhone = useRealPhonesData();
@@ -42,7 +35,6 @@ export const PhoneDashboard = () => {
   );
   const [showingPrimaryDashboard, toggleDashboardPanel] = useState(true);
   const dateToFormat = phoneDateCreated.data?.[0].verified_date!;
-  const inboundArray = inboundContactData.data;
 
   const toggleForwarding = () => {
     if (relayNumberData?.id) {
@@ -216,152 +208,21 @@ export const PhoneDashboard = () => {
     </div>
   );
 
-  const calendarStrings = {
-    //TODO: Add eng strings to pendingTranslations.ftl
-    lastDay: "[Yesterday at] LT",
-    sameDay: "[Today at] LT",
-    lastWeek: "L LT",
-    nextWeek: "L LT",
-    sameElse: "L LT",
-  };
-
-  const disabledCallerSMSSendersPanel = (
-    <div className={styles["disabled-senders-panel"]}>
-      <img
-        src={disabledSendersDataIllustration.src}
-        alt="Disabled Senders Data Illustration"
-        width={170}
-      />
-      <p className={styles["disabled-senders-panel-body"]}>
-        <WarningFilledIcon
-          alt=""
-          className={styles["warning-icon"]}
-          width={20}
-          height={20}
-        />
-        {l10n.getString("phone-dashboard-sender-disabled-body")}
-      </p>
-      <div className={styles["update-settings-cta"]}>
-        <OutboundLink
-          to="/accounts/settings/"
-          eventLabel="Update Settings"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          {l10n.getString("phone-dashboard-sender-disabled-update-settings")}
-        </OutboundLink>
-      </div>
-    </div>
-  );
-
-  const inboundContactArray =
-    inboundContactData &&
-    inboundArray
-      ?.sort(
-        (a, b) =>
-          // Sort by last sent date
-          moment(b.last_inbound_date).unix() -
-          moment(a.last_inbound_date).unix()
-      )
-      .map((data) => {
-        return (
-          <li
-            key={data.id}
-            className={data.blocked ? styles["greyed-contact"] : ""}
-          >
-            <span className={styles["sender-number"]}>
-              {formatPhoneNumberToUSDisplay(data.inbound_number)}
-            </span>
-            <span
-              className={`${styles["sender-date"]} ${styles["sender-date-wrapper"]}`}
-            >
-              {data.last_inbound_type === "text" && (
-                <ForwardedTextIcon
-                  alt="Last received a text"
-                  className={styles["forwarded-type-icon"]}
-                  width={20}
-                  height={12}
-                />
-              )}
-              {data.last_inbound_type === "call" && (
-                <ForwardedCallIcon
-                  alt="Last received a call"
-                  className={styles["forwarded-type-icon"]}
-                  width={20}
-                  height={15}
-                />
-              )}
-              <Moment calendar={calendarStrings}>
-                {data.last_inbound_date}
-              </Moment>
-            </span>
-            <span className={styles["sender-controls"]}>
-              <button
-                onClick={() =>
-                  inboundContactData.setForwardingState(!data.blocked, data.id)
-                }
-                className={styles["block-btn"]}
-              >
-                {data.blocked ? "Unblock" : "Block"}
-              </button>
-            </span>
-          </li>
-        );
-      });
-
-  const callerSMSSendersPanel = (
-    <div id="secondary-panel" className={styles["dashboard-card"]}>
-      <div className={styles["dashboard-card-caller-sms-senders-header"]}>
-        <span>
-          <button
-            type="button"
-            onClick={toggleSendersPanel}
-            className={styles["caller-sms-logs-back-btn"]}
-          >
-            <ChevronLeftIcon
-              alt="Back to Primary Dashboard"
-              className={styles["nav-icon"]}
-              width={20}
-              height={20}
-            />
-          </button>
-        </span>
-        <span className={styles["caller-sms-logs-title"]}>
-          {l10n.getString("phone-dashboard-senders-header")}
-        </span>
-        <span></span>
-      </div>
-
-      {/* Show senders UI only if enabled allows Caller and SMS Senders data in their settings */}
-      {profileData.data?.[0].store_phone_log ? (
-        <ul className={styles["caller-sms-senders-table"]}>
-          <li className={styles["greyed-contact"]}>
-            <span>
-              {l10n.getString("phone-dashboard-sender-table-title-sender")}
-            </span>
-            <span>
-              {l10n.getString("phone-dashboard-sender-table-title-activity")}
-            </span>
-            <span>
-              {l10n.getString("phone-dashboard-sender-table-title-action")}
-            </span>
-          </li>
-          {inboundContactArray}
-        </ul>
-      ) : (
-        <>{disabledCallerSMSSendersPanel}</>
-      )}
-    </div>
-  );
-
   return (
     <main className={styles["main-phone-wrapper"]}>
-      {showingPrimaryDashboard ? (
+      {showingPrimaryDashboard && inboundContactData !== undefined ? (
         // Primary Panel
         <>{primaryPanel}</>
       ) : (
         // Caller and SMS Senders Panel
-        <>{callerSMSSendersPanel}</>
+        <SendersPanelView
+          type="primary"
+          back_btn={{
+            onClick: () => {
+              toggleDashboardPanel;
+            },
+          }}
+        />
       )}
     </main>
   );
